Laraval queue:work doc updates
This adds the list of currently known queues to the worker job,
and also suggests not to use it. People who DO want to do it
themselves are left to burn their own bridges and make their
own mistakes.

This also added a few spaces and linting according to
vscode's interpretation of what valid Markdown is, and
fixes an accidental trailing backtick in the nginx section.

Signed-Off-By: Rob Thomas <xrobau@gmail.com>